--- plugin_daemonize_unix_test.go
@@ -1,11 +1,0 @@
-// +build !windows
-
-package plugin
-
-import (
-	"syscall"
-)
-
-func daemonize() {
-	syscall.Umask(0)
-}
