--- plugin_daemonize_windows_test.go
@@ -1,7 +1,0 @@
-// +build windows
-
-package plugin
-
-func daemonize() {
-	panic("Not supported")
-}
