+++ plugin_test.go
@@ -147,9 +147,6 @@ func TestHelperProcess(*testing.T) {
 		// Shouldn't reach here but make sure we exit anyways
 		os.Exit(0)
 	case "test-interface-daemon":
-		// Daemonize
-		daemonize()
-
 		// Serve!
 		Serve(&ServeConfig{
 			HandshakeConfig: testHandshake,
